+++ src/net_processing.h
@@ -61,7 +61,7 @@ public:
     * @param[in]   interrupt       Interrupt condition for processing threads
     * @return                      True if there is more work to be done
     */
-    bool SendMessages(CNode* pto, int node_index, int total_nodes, std::atomic<bool>& interrupt) override;
+    bool SendMessages(CNode* pto, size_t node_index, size_t total_nodes, std::atomic<bool>& interrupt) override;
 
     void ConsiderEviction(CNode *pto, int64_t time_in_seconds);
     void CheckForStaleTipAndEvictPeers(const Consensus::Params &consensusParams);
